Moved plugin-utils into a file in defs module.
commit-id:e52aa69a
