+++ crates/cairo-lang-defs/src/lib.rs
@@ -6,5 +6,6 @@ pub mod diagnostic_utils;
 pub mod ids;
 pub mod patcher;
 pub mod plugin;
+pub mod plugin_utils;
 #[cfg(test)]
 mod test;
