+++ crates/cairo-lang-semantic/src/inline_macros/assert.rs
@@ -2,12 +2,13 @@ use cairo_lang_defs::patcher::{PatchBuilder, RewriteNode};
 use cairo_lang_defs::plugin::{
     InlineMacroExprPlugin, InlinePluginResult, NamedPlugin, PluginDiagnostic, PluginGeneratedFile,
 };
+use cairo_lang_defs::plugin_utils::{
+    escape_node, try_extract_unnamed_arg, unsupported_bracket_diagnostic,
+};
 use cairo_lang_syntax::node::ast::WrappedArgList;
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, TypedStablePtr, TypedSyntaxNode};
 use indoc::formatdoc;
-
-use super::{escape_node, try_extract_unnamed_arg, unsupported_bracket_diagnostic};
 
 /// Macro for assertion.
 #[derive(Default, Debug)]
