+++ crates/cairo-lang-semantic/src/inline_macros/consteval_int.rs
@@ -1,3 +1,4 @@
+use cairo_lang_defs::extract_macro_single_unnamed_arg;
 use cairo_lang_defs::plugin::{
     InlineMacroExprPlugin, InlinePluginResult, NamedPlugin, PluginDiagnostic, PluginGeneratedFile,
 };
@@ -6,8 +7,6 @@ use cairo_lang_filesystem::span::{TextOffset, TextSpan, TextWidth};
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, TypedStablePtr, TypedSyntaxNode};
 use num_bigint::BigInt;
-
-use crate::extract_macro_single_unnamed_arg;
 
 #[derive(Debug, Default)]
 pub struct ConstevalIntMacro;
