+++ crates/cairo-lang-semantic/src/inline_macros/write.rs
@@ -4,13 +4,12 @@ use cairo_lang_defs::patcher::{PatchBuilder, RewriteNode};
 use cairo_lang_defs::plugin::{
     InlineMacroExprPlugin, InlinePluginResult, NamedPlugin, PluginDiagnostic, PluginGeneratedFile,
 };
+use cairo_lang_defs::plugin_utils::{try_extract_unnamed_arg, unsupported_bracket_diagnostic};
 use cairo_lang_filesystem::span::{TextSpan, TextWidth};
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, TypedStablePtr, TypedSyntaxNode};
 use cairo_lang_utils::{try_extract_matches, OptionHelper};
 use num_bigint::{BigInt, Sign};
-
-use super::{try_extract_unnamed_arg, unsupported_bracket_diagnostic};
 
 /// Macro for writing into a formatter.
 #[derive(Debug, Default)]
