+++ crates/cairo-lang-starknet/src/inline_macros/get_dep_component.rs
@@ -1,8 +1,8 @@
+use cairo_lang_defs::extract_macro_unnamed_args;
 use cairo_lang_defs::patcher::{PatchBuilder, RewriteNode};
 use cairo_lang_defs::plugin::{
     InlineMacroExprPlugin, InlinePluginResult, NamedPlugin, PluginDiagnostic, PluginGeneratedFile,
 };
-use cairo_lang_semantic::extract_macro_unnamed_args;
 use cairo_lang_syntax::node::db::SyntaxGroup;
 use cairo_lang_syntax::node::{ast, TypedStablePtr, TypedSyntaxNode};
 use cairo_lang_utils::extract_matches;
